requirements.txt
flask
python-dotenv
gunicorn
requests
duckdb
pandas
openai
google-generativeai
flask-cors
streamlit
